+++ glamod_marine_processing/cli_obs.py
@@ -25,6 +25,9 @@ def ObsCli(
     release,
     update,
     dataset,
+    process_list,
+    year_init,
+    year_end,
     source_pattern,
     prev_file_id,
     data_directory,
@@ -71,6 +74,11 @@ def ObsCli(
     level_config["run_jobs"] = run_jobs
     level_config["level"] = level
     level_config["overwrite"] = overwrite
+    if isinstance(process_list, str):
+        process_list = [process_list]
+    level_config["process_list"] = process_list
+    level_config["year_init"] = year_init
+    level_config["year_end"] = year_end
     if source_pattern:
         level_config["source_pattern"] = source_pattern
     if prev_file_id: