+++ tests/_testing_suite.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import os
 
 import _load_data
-import _settings
 import pandas as pd
+from _settings import get_settings
 from cdm_reader_mapper.cdm_mapper import read_tables
 from cdm_reader_mapper.common.getting_files import load_file
 
@@ -18,33 +18,40 @@ add_data = {
 }
 
 
-def manipulate_expected(expected, level):
-    """Manipulate expected result data."""
-    if level in _settings.manipulation.keys():
-        for index, values in _settings.manipulation[level].items():
-            expected[index] = values
-    if level in _settings.drops.keys():
-        expected = expected.drop(_settings.drops[level]).reset_index(drop=True)
-    return expected
-
-
-def _obs_testing(level, capsys):
+def _obs_testing(dataset, level, capsys):
     """Observational testing suite."""
+
+    def manipulate_expected(expected, level):
+        """Manipulate expected result data."""
+        if not hasattr(_settings, "manipulation"):
+            return expected
+        if level in _settings.manipulation.keys():
+            for index, values in _settings.manipulation[level].items():
+                expected[index] = values
+        if level in _settings.drops.keys():
+            expected = expected.drop(_settings.drops[level]).reset_index(drop=True)
+        return expected
+
+    _settings = get_settings(dataset)
     tables = _settings.which_tables[level]
     if add_data[level] is not None:
         add_data[level](
             cache_dir=f"./T{level}/release_7.0",
         )
 
-    cache_dir = _load_data.load_input(level)
+    cache_dir = _load_data.load_input(dataset, level, _settings)
 
     s = (
         "obs_suite "
         f"-l {level} "
+        f"-d {dataset} "
         f"-data_dir ./T{level} "
         f"-work_dir ./T{level} "
         f"-sp {_settings.pattern[level]} "
-        "-p_id subset "
+        f"-p_id subset "
+        f"-year_i {_settings.year_init} "
+        f"-year_e {_settings.year_end} "
+        f"-p_list {_settings.process_list} "
         "-o "
         "-run"
     )
@@ -53,16 +60,18 @@ def _obs_testing(level, capsys):
     assert captured.out == ""
 
     results = read_tables(
-        f"./T{level}/release_7.0/ICOADS_R3.0.2T/{level}/114-992", cdm_subset=tables
+        f"./T{level}/release_7.0/{dataset}/{level}/{_settings.deck}", cdm_subset=tables
     )
+
     for table_name in tables:
         load_file(
-            f"imma1_992/cdm_tables/{table_name}-114-992_2022-01-01_subset.psv",
-            cache_dir=f"./E{level}/ICOADS_R3.0.2T/{level}/114-992",
+            f"{_settings.input_dir}/cdm_tables/{table_name}-{_settings.cdm}.psv",
+            cache_dir=f"./E{level}/{dataset}/{level}/{_settings.deck}",
             within_drs=False,
         )
+
     expected = read_tables(
-        f"./E{level}/ICOADS_R3.0.2T/{level}/114-992", cdm_subset=tables
+        f"./E{level}/{dataset}/{level}/{_settings.deck}", cdm_subset=tables
     )
 
     expected = manipulate_expected(expected, level)